+++ src/main/index.ts
@@ -50,87 +50,57 @@ function getDigit_1111(number: number): string {
 
   return char1000 + char0111;
 }
-function getDigit_11111111(number: number): string {
+
+function getDigit_1111w(number: number): string {
   const digit10000 = (number / 10000) | 0;
   const digit01111 = number % 10000 | 0;
 
-  let char10000 = '';
-  if (digit10000 > 0 && digit10000 < 10) char10000 = getDigit_1(digit10000);
-  else if (digit10000 >= 10 && digit10000 < 100)
-    char10000 = getDigit_11(digit10000);
-  else if (digit10000 >= 100 && digit10000 < 1000)
-    char10000 = getDigit_111(digit10000);
-  else if (digit10000 >= 1000) char10000 = getDigit_1111(digit10000);
-
-  char10000 += CHARACTER_SET.TENTHOUSAND;
+  const char10000 = getDigit_9999(digit10000) + CHARACTER_SET.TENTHOUSAND;
 
   let char01111 = '';
-  if (digit01111 > 0 && digit01111 < 10)
-    char01111 = CHARACTER_SET.ZERO + getDigit_1(digit01111);
-  else if (digit01111 >= 10 && digit01111 < 100)
-    char01111 = CHARACTER_SET.ZERO + getDigit_11(digit01111);
-  else if (digit01111 >= 100 && digit01111 < 1000)
-    char01111 = CHARACTER_SET.ZERO + getDigit_111(digit01111);
-  else if (digit01111 >= 1000) char01111 = getDigit_1111(digit01111);
+  if (digit01111 > 0) char01111 = getDigit_9999(digit01111);
+  if (digit01111 > 0 && digit01111 < 1000)
+    char01111 = CHARACTER_SET.ZERO + char01111;
 
   return char10000 + char01111;
 }
-function getDigit_111122223333(number: number): string {
-  const YI = 100000000;
+function getDigit_1111yi(number: number): string {
+  const digit10000w = (number / 100000000) | 0;
+  const digit01111w = number % 100000000 | 0;
 
-  const digit100000000 = (number / YI) | 0;
-  const digit011111111 = number % YI | 0;
+  let char10000w = getDigit_9999(digit10000w) + CHARACTER_SET.HUNDREDMILLION;
 
-  let char100000000 = '';
-  if (digit100000000 > 0 && digit100000000 < 10)
-    char100000000 = getDigit_1(digit100000000);
-  else if (digit100000000 >= 10 && digit100000000 < 100)
-    char100000000 = getDigit_11(digit100000000);
-  else if (digit100000000 >= 100 && digit100000000 < 1000)
-    char100000000 = getDigit_111(digit100000000);
-  else if (digit100000000 >= 1000)
-    char100000000 = getDigit_1111(digit100000000);
-
-  char100000000 += CHARACTER_SET.HUNDREDMILLION;
-
-  const WAN = 10000;
-  const digit10000 = (digit011111111 / WAN) | 0;
-  const digit01111 = digit011111111 % WAN | 0;
+  const digit10000 = (digit01111w / 10000) | 0;
+  const digit01111 = digit01111w % 10000 | 0;
 
   let char10000 = '';
-  if (digit10000 > 0 && digit10000 < 10) char10000 = getDigit_1(digit10000);
-  else if (digit10000 >= 10 && digit10000 < 100)
-    char10000 = getDigit_11(digit10000);
-  else if (digit10000 >= 100 && digit10000 < 1000)
-    char10000 = getDigit_111(digit10000);
-  else if (digit10000 >= 1000) char10000 = getDigit_1111(digit10000);
-
-  if (digit10000 > 0) char10000 += CHARACTER_SET.TENTHOUSAND;
-
-  if (digit10000 < 1000 && digit10000 > 0) char100000000 += CHARACTER_SET.ZERO;
+  if (digit10000 > 0) {
+    char10000 = getDigit_9999(digit10000) + CHARACTER_SET.TENTHOUSAND;
+    if (digit10000 < 1000) char10000w += CHARACTER_SET.ZERO;
+  }
 
   let char01111 = '';
-  if (digit01111 > 0 && digit01111 < 10)
-    char01111 = CHARACTER_SET.ZERO + getDigit_1(digit01111);
-  else if (digit01111 >= 10 && digit01111 < 100)
-    char01111 = CHARACTER_SET.ZERO + getDigit_11(digit01111);
-  else if (digit01111 >= 100 && digit01111 < 1000)
-    char01111 = CHARACTER_SET.ZERO + getDigit_111(digit01111);
-  else if (digit01111 >= 1000 && digit10000 === 0)
-    char01111 = CHARACTER_SET.ZERO + getDigit_1111(digit01111);
-  else if (digit01111 >= 1000 && digit10000 > 0)
-    char01111 = getDigit_1111(digit01111);
+  if (digit01111 > 0) {
+    char01111 = getDigit_9999(digit01111);
 
-  return char100000000 + char10000 + char01111;
+    if (digit01111 < 1000 || (digit01111 >= 1000 && digit10000 === 0))
+      char01111 = CHARACTER_SET.ZERO + char01111;
+  }
+
+  return char10000w + char10000 + char01111;
 }
+function getDigit_9999(number: number): string {
+  if (number < 10) return getDigit_1(number);
+  else if (number < 100) return getDigit_11(number);
+  else if (number < 1000) return getDigit_111(number);
+  else return getDigit_1111(number);
+}
+
 export function convertNumber(number: number): string {
   if (number < 0 || number > 999999999999) throw ERR_NOT_SUPPORTED;
   if (number !== Math.floor(number)) throw ERR_NOT_SUPPORTED;
 
-  if (number < 10) return getDigit_1(number);
-  else if (number < 100) return getDigit_11(number);
-  else if (number < 1000) return getDigit_111(number);
-  else if (number < 10000) return getDigit_1111(number);
-  else if (number < 100000000) return getDigit_11111111(number);
-  return getDigit_111122223333(number);
+  if (number < 10000) return getDigit_9999(number);
+  else if (number < 100000000) return getDigit_1111w(number);
+  return getDigit_1111yi(number);
 }